requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
flask>=3.0.0
flask-cors>=4.0.0
anthropic>=0.49.0
joblib>=1.3.0
matplotlib>=3.7.0
plotly>=5.18.0
python-dotenv>=1.0.0
